let's not introduce ROOT_BUILD
This option is for legacy shell scripts, and probably was only needed
for ease of cleanup. In Snakemake world there are different ways to
manage those files.

Enabling this would require us to introduce use of the
compile_analysis.py for other existing detector_benchmarks such as
hcal_barrel and zdc.

We should aim to deprecate ROOT_BUILD for physics_benchmarks instead.